requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.25
pydantic-settings==2.2.1
python-multipart==0.0.9
pytest==8.0.0
httpx==0.27.0
pandas==2.2.2
pydantic==2.6.1
python-dotenv==1.0.1